+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+ - Fixed non-clickable links in tables.
+
 ## [2.3.1]
 
-- Adds text formatting ability for paragraph types: "Image with caption (thumbnail)" and "List"
+ - Adds text formatting ability for paragraph types: "Image with caption (thumbnail)" and "List"
 
 ## [2.3.0]